k8s: move service related functions into service.go
To reduce the size and scope of the file k8s/watchers/watcher.go, this
commit moves all service relevant functions into the file service.go.
